Rename model to dto
